make PPMStatusTimeline a connected component
