+++ pkg/auth/authentication/devlocal.go
@@ -54,7 +54,7 @@ func (h UserListHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		http.Redirect(w, r, h.landingURL(session), http.StatusTemporaryRedirect)
 		return
 	}
-	limit := 25
+	limit := 100
 	identities, err := models.FetchAppUserIdentities(h.db, session.ApplicationName, limit)
 	if err != nil {
 		h.logger.Error("Could not load list of users", zap.Error(err))
